requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
mutagen==1.46.0
pycparser==2.21
pycryptodomex==3.15.0
PyNaCl==1.5.0
websockets==10.4
yarl==1.8.1
yt-dlp==2022.10.4
